formal: refactor and unify formality checks
Formality checks are now unified into a single script that is meant to
be shared between all the feeds.

Author and Signed-off-by email cannot be a GitHub noreply email as per
the contributing guidelines.

Add Weblate exceptions for the luci feed.

Extract the script to a separate file for easier maintainability and
debugging.

Signed-off-by: George Sapkin <george@sapk.in>